requirements.txt
#main
boto3==1.26.100
celery==5.2.7
crispy-bootstrap4==2022.1
Django== 4.1.6
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.1
django-debug-toolbar==3.8.1
django-admin-rangefilter==0.9.0
django-import-export==3.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-yasg==1.21.5
ipython==8.10.0
requests==2.28.2

#services
gunicorn==20.1.0
psycopg2-binary==2.9.6
pymemcache==4.0.0


#flake8
flake8==6.0.0
flake8-builtins==2.1.0
flake8-print==5.0.0

#pytest
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0

